requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==38.0.3
cycler==0.11.0
fonttools==4.38.0
greenlet==2.0.1
importlib-metadata==5.0.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
paramiko==2.12.0
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
seaborn==0.12.1
six==1.16.0
SQLAlchemy==1.4.44
typing_extensions==4.4.0
zipp==3.10.0
python-decouple==3.6
python-decouple 3.6
djangorestframework==3.14.0
